+++ serialisation/src/test/java/org/ehrbase/openehr/sdk/serialisation/flatencoding/FlatTestHelper.java
@@ -1,0 +1,51 @@
+/*
+ * Copyright (c) 2024 vitasystems GmbH and Hannover Medical School.
+ *
+ * This file is part of project openEHR_SDK
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.ehrbase.openehr.sdk.serialisation.flatencoding;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import org.ehrbase.openehr.sdk.serialisation.jsonencoding.ArchieObjectMapperProvider;
+
+public class FlatTestHelper {
+
+    @SuppressWarnings("unchecked")
+    public static List<String> compere(String actualJson, String expectedJson) throws JsonProcessingException {
+        List<String> errors = new ArrayList<>();
+        ObjectMapper objectMapper = ArchieObjectMapperProvider.getObjectMapper();
+
+        Map<String, Object> actual = objectMapper.readValue(actualJson, Map.class);
+        Map<String, Object> expected = objectMapper.readValue(expectedJson, Map.class);
+
+        actual.forEach((key, value) -> {
+            if (!expected.containsKey(key) || !expected.get(key).equals(value)) {
+                errors.add(String.format("Missing path: %s, value: %s", key, value));
+            }
+        });
+
+        expected.forEach((key, value) -> {
+            if (!actual.containsKey(key) || !actual.get(key).equals(value)) {
+                errors.add(String.format("Extra path: %s, value: %s", key, value));
+            }
+        });
+
+        return errors;
+    }
+}